gh-72 Moved the posts into a seperate component and then rendered them on the home page
